# backend/openapi.yaml
openapi: 3.0.0
info:
  title: QQM API
  version: 1.0.0
  description: Client / project / infrastructure management API. Generated from
    the actual implemented routes, controllers, and zod validators
    (backend/src/openapi) — not from any prior design document.
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: /
    description: Same-origin backend
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT obtained from POST /api/auth/login, sent as `Authorization:
        Bearer <token>`."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - message
      required:
        - error
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
    ClientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          nullable: true
          format: date-time
      required:
        - id
        - name
        - status
        - description
        - created_at
        - updated_at
        - deleted_at
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required:
        - page
        - per_page
        - total
        - total_pages
...
    PeopleType:
      type: string
      enum:
        - internal_engineer
        - vendor
        - client_contact
        - project_owner
        - approver
...
    PersonListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/PeopleType"
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          nullable: true
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - type
        - notes
        - created_at
        - updated_at
        - deleted_at
    PersonDetail:
      allOf:
        - $ref: "#/components/schemas/Person"
        - type: object
          properties:
            clients:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  relationship_type:
                    type: string
                    nullable: true
                required:
                  - id
                  - name
                  - relationship_type
            account:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                  format: uuid
                username:
                  type: string
                active:
                  type: boolean
              required:
                - id
                - username
                - active
          required:
            - clients
            - account
...
paths:
...
  /api/clients:
    get:
      tags:
        - Clients
      operationId: listClients
      summary: List clients
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            description: "Default: 1"
          required: false
          description: "Default: 1"
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            description: "Default: 20, max: 100"
          required: false
          description: "Default: 20, max: 100"
          name: per_page
          in: query
        - schema:
            type: string
            enum:
              - name
              - status
              - created_at
              - updated_at
            description: "Default: name"
          required: false
          description: "Default: name"
          name: sort
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            description: "Default: asc"
          required: false
          description: "Default: asc"
          name: order
          in: query
        - schema:
            type: string
            enum:
              - "false"
              - "true"
              - all
            description: "Default: false"
          required: false
          description: "Default: false"
          name: deleted
          in: query
        - schema:
            type: string
          required: false
          name: search
          in: query
      responses:
        "200":
          description: Paginated clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Clients
      operationId: createClient
      summary: Create a client (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                status:
                  type: string
                description:
                  type: string
              required:
                - name
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/people:
    get:
      tags:
        - People
      operationId: listPeople
      summary: List people
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            description: "Default: 1"
          required: false
          description: "Default: 1"
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            description: "Default: 20, max: 100"
          required: false
          description: "Default: 20, max: 100"
          name: per_page
          in: query
        - schema:
            type: string
            enum:
              - name
              - created_at
              - updated_at
            description: "Default: name"
          required: false
          description: "Default: name"
          name: sort
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            description: "Default: asc"
          required: false
          description: "Default: asc"
          name: order
          in: query
        - schema:
            type: string
            enum:
              - "false"
              - "true"
              - all
            description: "Default: false"
          required: false
          description: "Default: false"
          name: deleted
          in: query
        - schema:
            type: string
          required: false
          name: search
          in: query
        - schema:
            type: string
          required: false
          name: type
          in: query
      responses:
        "200":
          description: Paginated people
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - People
      operationId: createPerson
      summary: Create a person (admin or member)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                phone:
                  type: string
                type:
                  type: string
                  enum:
                    - internal_engineer
                    - vendor
                    - client_contact
                    - project_owner
                    - approver
                notes:
                  type: string
              required:
                - name
                - type
      responses:
        "201":
          description: Person created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/people/{id}:
    get:
      tags:
        - People
      operationId: getPersonById
      summary: Get a person by id, with linked clients and (if any) their user account
        inlined
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Person
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - People
      operationId: updatePerson
      summary: Update a person (admin or member; optimistic lock)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                phone:
                  type: string
                type:
                  type: string
                  enum:
                    - internal_engineer
                    - vendor
                    - client_contact
                    - project_owner
                    - approver
                notes:
                  type: string
                updated_at:
                  type: string
                  minLength: 1
              required:
                - updated_at
      responses:
        "200":
          description: Person updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - People
      operationId: deletePerson
      summary: Soft-delete a person (admin only; also disables their linked user
        account, if any, blocking login)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Person soft-deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/people/{id}/restore:
    post:
      tags:
        - People
      operationId: restorePerson
      summary: Restore a soft-deleted person (admin only; also re-enables their linked
        user account, if it was disabled)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Person restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...